requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.4.0.2
bcrypt==3.1.4
bravado==9.2.2
bravado-core==4.11.2
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
enum34==1.1.6
functools32==3.2.3.post2
idna==2.6
iniparse==0.4
ipaddress==1.0.19
jsonref==0.1
jsonschema==2.6.0
kitchen==1.1.1
msgpack-python==0.5.1
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
PyNaCl==1.2.1
python-dateutil==2.6.1
pytz==2017.3
pyvmomi==6.5.0.2017.5.post1
pyxattr==0.5.1
PyYAML==3.12
requests==2.18.4
rfc3987==1.3.7
simplejson==3.13.2
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
tabulate==0.8.2
urlgrabber==3.10
urllib3==1.22
webcolors==1.7
yum-metadata-parser==1.1.4
